refactor calendar-event-import-manager
